move casts

+++ src/main.rs
@@ -16,17 +16,17 @@ fn make_grid() -> Grid {
 
 /// Euclidean modulo
 /// As long as n is positive, the result will always be positive
-fn modulo(a: isize, n: isize) -> isize {
+fn modulo(a: isize, n: isize) -> usize {
     if a < 0 {
-        (a % n + n) % n
+        ((a % n + n) % n) as usize
     } else {
-        a % n
+        (a % n) as usize
     }
 }
 
 /// Count the number of neighbouring cells that are alive, not including itself
 /// Wraps around the grid edges, effectively making it a torus
-fn alive_neighbours(grid: &Grid, x: isize, y: isize) -> u8 {
+fn alive_neighbours(grid: &Grid, x: usize, y: usize) -> u8 {
     let mut count = 0;
     let offsets = [
         (-1, -1),
@@ -39,8 +39,8 @@ fn alive_neighbours(grid: &Grid, x: isize, y: isize) -> u8 {
         (1, 1),
     ];
     for (x_o, y_o) in &offsets {
-        let xindex = modulo(x_o + x, WIDTH as isize) as usize;
-        let yindex = modulo(y_o + y, HEIGHT as isize) as usize;
+        let xindex = modulo(x_o + x as isize, WIDTH as isize);
+        let yindex = modulo(y_o + y as isize, HEIGHT as isize);
 
         count += grid[yindex][xindex] as u8;
     }
@@ -51,7 +51,7 @@ fn alive_neighbours(grid: &Grid, x: isize, y: isize) -> u8 {
 fn generation(grid: &Grid, next: &mut Grid) {
     for (y, row) in grid.iter().enumerate() {
         for (x, alive) in row.iter().enumerate() {
-            next[y][x] = match alive_neighbours(&grid, x as isize, y as isize) {
+            next[y][x] = match alive_neighbours(&grid, x, y) {
                 2 => *alive,
                 3 => true,
                 _ => false,
